% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_detect.R
\name{outlier_detect}
\alias{outlier_detect}
\title{Detects the Presence of Outlier(s) in all the Columns of Data}
\usage{
outlier_detect(data)
}
\arguments{
\item{data}{}
}
\value{
matrix of column names and status of outliers
}
\description{
This package helps in detecting the presence of outliers in data. It returns a matrix of column names and whether the outlier was detected in it or not.
}
\examples{
outlier_detect(Boston)

}
