% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_numbers.R
\name{outlier_numbers}
\alias{outlier_numbers}
\title{Counts the No. of Outliers Before and After Log Transformation of Variables.}
\usage{
outlier_numbers(data)
}
\arguments{
\item{data_frame}{}
}
\value{
matrix with column names and no. of outliers.
}
\description{
This package accepts a data set as an argument and returns the outlier counts for each variables in the data set.
}
\examples{
outlier_numbers(Boston)

}
