% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmissing.R
\name{pmissing}
\alias{pmissing}
\title{Calculates Percentage of Missing Values}
\usage{
pmissing(data)
}
\arguments{
\item{data}{}
}
\value{
matrix of column names and the percentage of missing values
}
\description{
This package calculates and returns a matrix of columns of a data frame and the percentage of missing values in it.
}
\examples{
pmissing(Boston)

}
